adding a new freecad file with sketch on the right face.
not sure how to remove the disassociated bodies extruded from the
imported svg logo from the previous attempt. Now I've mapped the sketch
on the face of the overmold, but I'm not sure how to move the sketch. I
get this weird wirefame of green spiraly lines that cover-up everything
on the screen when I attempt to edit the sketch..