next attempt at getting the logo pocketed from the overmold
This time I learned how to use the draft tool to rotate & move a sketch,
then I also made the individual sketches into distinct bodies in a
group. And I dropped their location on the y axis to sit just on-top of
the overmold.

Next I just need to figure out how to pad them and then subtrack them
from the overmold, which seems non-trivial because the imported overmod
is not a proper body or something..